Migrate to modern ts architecture and es modules.
