+++ .changeset/full-needles-relate.md
@@ -1,0 +1,5 @@
+---
+"@irsdk-node/native": minor
+---
+
+Removed redundant manual type declaration of native SDK class.
