+++ .changeset/slimy-dingos-repeat.md
@@ -1,0 +1,7 @@
+---
+"@irsdk-node/native": minor
+"@irsdk-node/types": minor
+"irsdk-node": minor
+---
+
+Migrated codebase to modern typescript and es-modules
