+++ packages/irsdk-node-native/.vscode/settings.json
@@ -6,7 +6,7 @@
     "**/CVS": true,
     "**/.DS_Store": true,
     "**/Thumbs.db": true,
-    "**/build": true,
+    "**/build": false,
     "**/dist": false,
     "**/node_modules": true,
   },
