+++ packages/irsdk-node-native/package.json
@@ -24,15 +24,21 @@
     "./binding.gyp",
     "./README.md"
   ],
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "default": "./dist/index.js"
+    }
+  },
   "main": "./dist/index.js",
-  "types": "./src/index.d.ts",
+  "types": "./dist/index.d.ts",
   "scripts": {
     "clean:ts": "rm -rf dist *.tsbuildinfo",
     "clean:cpp": "node-gyp clean",
     "clean": "pnpm run \"/^clean:.*/\"",
     "configure": "pnpm run clean:cpp && node-gyp configure",
     "build": "pnpm run configure && pnpm run \"/^build:(ts|cpp)$/\"",
-    "build:ts": "tsc --build --force",
+    "build:ts": "tsc",
     "build:cpp": "pnpm run \"/^build:cpp\\-.*/\"",
     "build:cpp-release": "node-gyp build",
     "build:cpp-debug": "node-gyp build --debug",
