+++ packages/irsdk-node-native/src/globals.d.ts
@@ -1,0 +1,5 @@
+/* eslint-disable @typescript-eslint/unified-signatures */
+
+declare module '*.node' {
+  export const iRacingSdkNode: unknown;
+}
