+++ packages/irsdk-node-types/src/index.ts
@@ -1,10 +1,10 @@
-export * from './enums';
-export type * from './camera-info';
-export type * from './driver-info';
-export type * from './radio-info';
-export type * from './session-info';
-export type * from './setup-info';
-export type * from './split-info';
-export type * from './weekend-info';
-export type * from './session-yaml';
-export type * from './_GENERATED_telemetry';
+export * from './enums.js';
+export type * from './camera-info.js';
+export type * from './driver-info.js';
+export type * from './radio-info.js';
+export type * from './session-info.js';
+export type * from './setup-info.js';
+export type * from './split-info.js';
+export type * from './weekend-info.js';
+export type * from './session-yaml.js';
+export type * from './_GENERATED_telemetry.js';
