+++ packages/irsdk-node-types/src/session-yaml.ts
@@ -1,10 +1,10 @@
-import { CameraInfo } from './camera-info';
-import { DriverInfo } from './driver-info';
-import { RadioInfo } from './radio-info';
-import { SessionList } from './session-info';
-import { CarSetupInfo } from './setup-info';
-import { SplitTimeInfo } from './split-info';
-import { WeekendInfo } from './weekend-info';
+import { CameraInfo } from './camera-info.js';
+import { DriverInfo } from './driver-info.js';
+import { RadioInfo } from './radio-info.js';
+import { SessionList } from './session-info.js';
+import { CarSetupInfo } from './setup-info.js';
+import { SplitTimeInfo } from './split-info.js';
+import { WeekendInfo } from './weekend-info.js';
 
 /**
  * Information about the current session, stored as yaml.
