+++ packages/irsdk-node-types/tsconfig.json
@@ -1,14 +1,11 @@
 {
   "extends": "../../tsconfig.base.json",
-  "references": [],
-  "include": [
-    "src/**/*.ts"
-  ],
+  "include": ["./src/**/*.ts",],
+  "exclude": ["dist"],
+  
   "compilerOptions": {
-    "composite": true,
-    "rootDir": "./src",
+    "tsBuildInfoFile": "./.tsbuildinfo/tsconfig.tsbuildinfo",
     "outDir": "./dist",
-    "baseUrl": "./src",
-    "declarationDir": "dist"
-  }
+    "rootDir": "./src",
+  },
 }
