+++ packages/irsdk-node/src/get-sdk.ts
@@ -1,9 +1,10 @@
-import { platform } from 'os';
+import { platform } from 'node:os';
 
 import type { INativeSDK } from '@irsdk-node/native';
 
-import { MockSDK } from './utils/mock-sdk';
+import { MockSDK } from './utils/mock-sdk.js';
 
+// TODO: Return mock from C++, not this nonsense.
 export async function getSdkOrMock(): Promise<INativeSDK> {
   if (platform() === 'win32') {
     const Sdk = (await import('@irsdk-node/native')).NativeSDK;
