+++ packages/irsdk-node/src/index.ts
@@ -1,1 +1,1 @@
-export * from './irsdk-node';
+export * from './irsdk-node.js';
