--- packages/irsdk-node/src/utils/index.ts
@@ -1,1 +1,0 @@
-export * from './sim-status';
