--- tsconfig.json
@@ -1,20 +1,0 @@
-{
-  "extends": "./tsconfig.base.json",
-  "include": [
-    "./packages/irsdk-node-native/src/**/*.ts",
-    "./packages/irsdk-node-types/src/**/*.ts",
-    "./packages/irsdk-node/src/**/*.ts",
-  ],
-  "references": [
-    { "path": "./packages/irsdk-node-native" },
-    { "path": "./packages/irsdk-node-types" },
-    { "path": "./packages/irsdk-node" },
-  ],
-  // "compilerOptions": {
-  //   "paths": {
-  //     "@irsdk-node/native": ["./packages/irsdk-node-native/src"],
-  //     "@irsdk-node/types": ["./packages/irsdk-node-types/src"],
-  //     "irsdk-node": ["./packages/irsdk-node/src"],
-  //   }
-  // }
-}
